User -> User, Account Comment 리팩토링
